+++ packages/optimizely-cms-sdk/src/model/contentTypeRegistry.ts
@@ -1,3 +1,4 @@
+import { CUSTOM_MEDIA_TYPE, normalizeBaseType } from '../util/baseTypeUtil';
 import { AnyContentType, MediaStringTypes } from './contentTypes';
 
 let _registry: AnyContentType[] = [];
@@ -17,34 +18,6 @@ export function getAllContentTypes(): AnyContentType[] {
   return _registry;
 }
 
-/** Normalize CMS baseType to SDK types */
-export function normalizeBaseType(key: string): string {
-  switch (key) {
-    case '_Image':
-      return 'image';
-    case '_Media':
-      return 'media';
-    case '_Video':
-      return 'video';
-    default:
-      return key;
-  }
-}
-
-/** get baseType from custom item */
-export function getBaseKey(key: string): string {
-  switch (key) {
-    case 'image':
-      return ' _Image';
-    case 'media':
-      return ' _Media';
-    case 'video':
-      return ' _Video';
-    default:
-      return key;
-  }
-}
-
 /** Get the Component from a base type */
 export function getContentTypeByBaseType(
   name: MediaStringTypes
@@ -57,6 +30,6 @@ export function getContentTypeByBaseType(
 /** Get the Component from a content type name */
 export function getAllMediaTypeKeys() {
   return _registry
-    .filter((c) => ['image', 'media', 'video'].includes(c.baseType))
+    .filter((c) => CUSTOM_MEDIA_TYPE.includes(c.baseType))
     .map((c) => c.key);
 }
