+++ packages/optimizely-cms-sdk/src/util/baseTypeUtil.ts
@@ -70,3 +70,31 @@ export function buildBaseTypeFragments(baseType: MediaStringTypes) {
   }
   return { fields: [], extraFragments: [] };
 }
+
+/** Normalize CMS baseType to SDK types */
+export function normalizeBaseType(key: string): string {
+  switch (key) {
+    case '_Image':
+      return 'image';
+    case '_Media':
+      return 'media';
+    case '_Video':
+      return 'video';
+    default:
+      return key;
+  }
+}
+
+/** get baseType from custom item */
+export function getBaseKey(key: string): string {
+  switch (key) {
+    case 'image':
+      return ' _Image';
+    case 'media':
+      return ' _Media';
+    case 'video':
+      return ' _Video';
+    default:
+      return key;
+  }
+}
